GTPLBankAutomation_UI/src/main/java/com/gtpl/qa/pages/LoginPage.java
package com.gtpl.qa.pages;

import org.openqa.selenium.WebElement;
import org.openqa.selenium.support.FindBy;
import org.openqa.selenium.support.PageFactory;

import com.gtpl.qa.base.GTPLBase;

public class LoginPage extends GTPLBase {
	
	//@FindBy()
	
	@FindBy(name="uid")
	private WebElement username;
	
	@FindBy(name="password")
	private WebElement password;
	
	@FindBy(name="btnLogin")
	private WebElement loginBtn;

	@FindBy(xpath="//h2[text()='Gtpl Bank']")
	private WebElement bankLogo;
	
	public LoginPage() {
		PageFactory.initElements(driver, this);
	}
	
	public String validateLoginPageTitle() {
		return driver.getTitle();
	}
	
	public boolean validateBankLogo() {
		return bankLogo.isDisplayed();
	}
	
	public HomePage login(String un,String pwd) {
		username.sendKeys(un);
		password.sendKeys(pwd);
		loginBtn.click();
		return new HomePage();
	}

}
